updated
update the structure
