+++ client/src/App.jsx
@@ -9,12 +9,10 @@ import ProfilePage from "./components/UserPortalUI/Profile/Profile";
 import MealsPage from "./components/UserPortalUI/Meals/Meals"; // Import MealsPage
 import SuggestionsPage from "./components/UserPortalUI/Suggestions/Suggestions"; // Import SuggestionsPage
 import AttendancePage from "./components/UserPortalUI/Attendance/Attendance"; // Import AttendancePage
-import KitchenAdmin from "./pages/KitchenAdmin/KitchenAdminDashBoard/kitchenAdminDash";
-import Reports from "./pages/KitchenAdmin/Report/report";
-import Meals from "./pages/KitchenAdmin/MealDash/meal";
+import KitchenAdmin from "./components/pages/Kitchen Admin/KitchenAdminDashBoard/kitchenAdminDash";
+import Reports from "./components/pages/Kitchen Admin/Report/report";
+import Meals from "./components/pages/Kitchen Admin/MealDash/meal";
 import MealDetailsForm from "./pages/KitchenAdmin/MealDetails/MealDetailsForm";
-import MealPlan from "./pages/KitchenAdmin/MealPlan/mealPlan";
-import SelectIngredients from "./pages/KitchenAdmin/SelectIngredients/ingredients";
 import EditMeal from "./pages/KitchenAdmin/EditMeal/editMeal";
 import Login from "./pages/Login/Login";
 import Serving from "./pages/Serving staff/Barcode Scan/Serving";
@@ -55,9 +53,7 @@ function App() {
           <Route path="/report" element={<Reports />} />
           <Route path="/kitchen-report" element={<Reports />} />
           <Route path="/kitchen-meal" element={<Meals />} />
-          <Route path="/meal-details" element={<MealDetailsForm />} />
-          <Route path="/meal-plan" element={<MealPlan />} />
-          <Route path="/select-ingredients" element={<SelectIngredients />} />
+          <Route path="/meal-details" element={<MealDetailsForm />} />        
           <Route path="/edit-meal" element={<EditMeal />} />
           <Route path="/serving" element={<Serving />} />
           <Route path="/meal-conform/:id" element={<MealConform />} />
