--- client/src/components/KitchenAdmin/Analysis/analysis.module.css
@@ -1,30 +1,0 @@
-.analysisContainer {
-    padding: 20px;
-    background: white;
-    border-radius: 8px;
-    border-radius: 8px;
-    box-shadow: 1px 0px 1px 2px rgba(0, 0, 0, 0.1);
-
-  }
-  
-  .heading {
-    font-size: 24px;
-    font-weight: bold;
-    margin-bottom: 20px;
-  }
-  
-  .cardsContainer {
-    display: flex;
-    gap: 20px;
-  }
-  
-  .card {
-    flex: 1;
-    border-radius: 8px;
-    box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
-  }
-  
-  .mealItem {
-    margin-bottom: 10px;
-  }
-
