--- client/src/components/KitchenAdmin/Ingredients/navbar.jsx
@@ -1,51 +1,0 @@
-
-import React from 'react';
-import { Button , Layout} from 'antd';
-import { ArrowLeftOutlined, LogoutOutlined } from '@ant-design/icons';
-import { useNavigate } from 'react-router-dom';
-
-const Nav = () => {
-  const navigate = useNavigate(); 
-
-  const handleDashboardClick = () => {
-    navigate('/meal-details');
-  };
-
-  const handleLogout = () => {
-    ocalStorage.removeItem('token'); 
-    localStorage.removeItem('user'); 
-    navigate('/login');
-  };
-
-  return (
-    <Layout.Header style={{ 
-      backgroundColor: '#800020', 
-      padding: '0 24px',
-      display: 'flex',
-      justifyContent: 'space-between',
-      alignItems: 'center',
-      color: 'white'
-    }}>
-      <Button
-        type="text"
-        icon={<ArrowLeftOutlined />}
-        style={{ color: 'white', padding: 0 }}
-        onClick={handleDashboardClick}
-      >
-        Back
-      </Button>
-      <Button
-        type="text"
-        icon={<LogoutOutlined />} 
-        style={{ color: 'white', padding: 0 }}
-        onClick={handleLogout}
-      >
-        Logout
-      </Button>
-    </Layout.Header>
-  );
-
-};
-
-
-export default Nav;
