--- client/src/components/KitchenAdmin/KitchenNavbar/navbar.module.css
@@ -1,297 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Inter:ital,opsz,wght@0,14..32,100..900;1,14..32,100..900&family=Itim&family=Roboto+Slab:wght@100..900&display=swap");
-@import url("https://fonts.googleapis.com/css2?family=Bree+Serif&family=Inter:ital,opsz,wght@0,14..32,100..900;1,14..32,100..900&family=Itim&family=Roboto+Slab:wght@100..900&display=swap");
-
-.sider {
-  background-color: rgb(244, 244, 244);
-  height: 100vh;
-}
-
-.userDropdown {
-  margin-right: 2vw;
-  position: relative;
-}
-
-.userInfo {
-  display: flex;
-  align-items: center;
-  gap: 12px;
-  padding: 8px 12px;
-  border-radius: 8px;
-  cursor: pointer;
-  transition: background-color 0.3s;
-}
-.userInfo:hover {
-  background-color: rgba(0, 0, 0, 0.04);
-}
-
-.userName {
-  font-weight: 600;
-  font-size: 18px;
-  /* font-family: 'Inter', sans-serif; */
-  color: #333;
-}
-.userPosition{
-  font-size: 14px;
-  color: #666;
-}
-
-/* Dropdown menu items styling */
-.profileMenuItem,
-.logoutMenuItem {
-  display: flex;
-  align-items: center;
-  padding: 10px 14px;
-  font-size: 15px;
-  font-weight: 400;
-  border-radius: 0;
-}
-
-.profileMenuItem {
-  color: #333;
-}
-
-.logoutMenuItem {
-  color: #f44336;
-}
-
-.menuItemIcon {
-  margin-right: 8px;
-  font-size: 16px;
-}
-
-.menu {
-  font-size: 0.9vw;
-  font-family: "Inter", sans-serif;
-  background-color: rgb(245, 244, 244);
-  font-weight: 500;
-}
-
-.navHeader {
-  text-align: center;
-  font-family: "Bree Serif", serif;
-  font-size: 1.6vw;
-  transition: opacity 0.3s ease-in-out;
-}
-.navHeader.collapsed {
-  opacity: 0;
-}
-
-.red {
-  color: #d10000;
-}
-
-.line {
-  width: 85%;
-  margin-bottom: 1vw;
-}
-
-.headerText {
-  margin: 43%;
-  font-size: 1.3vw;
-  font-weight: 650;
-  font-family: 4;
-}
-
-.headerContent {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  background-color: white;
-  position: relative;
-  z-index: 10;
-  box-shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.1),
-    0 2px 4px -1px rgba(0, 0, 0, 0.06);
-}
-
-img.logo {
-  margin: auto;
-  width: 18%;
-}
-
-.main {
-  display: flex;
-  height: 100vh;
-  overflow: hidden;
-}
-
-.homeContent {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  height: 100vh;
-  overflow: hidden;
-}
-
-.content {
-  overflow-y: auto !important;
-  height: 100%;
-  overflow-x: hidden;
-  flex: 1;
-}
-
-/* Notification menu item styling */
-.notificationMenuItem {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  width: 100%;
-}
-
-.notificationBadge {
-  background-color: #d10000;
-  color: white;
-  border-radius: 50%;
-  width: 20px;
-  height: 20px;
-  aspect-ratio: 1;
-  font-size: 12px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-left: auto;
-}
-
-/* Notification Panel */
-.notificationSidePanel {
-  position: fixed;
-  top: 0;
-  right: 0;
-  width: 350px;
-  height: 100vh;
-  background-color: white;
-  box-shadow: -2px 0 10px rgba(0, 0, 0, 0.1);
-  z-index: 1001;
-  display: flex;
-  flex-direction: column;
-  animation: slideIn 0.3s ease;
-  overflow: hidden;
-}
-
-@keyframes slideIn {
-  from {
-    transform: translateX(100%);
-  }
-  to {
-    transform: translateX(0);
-  }
-}
-
-.notificationHeader {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 15px 20px;
-  border-bottom: 1px solid #eee;
-  background-color: #fafafa;
-}
-
-.notificationHeader h2 {
-  margin: 0;
-  font-size: 18px;
-  font-weight: 600;
-  color: #333;
-}
-
-.notificationActions {
-  display: flex;
-  align-items: center;
-}
-
-.notificationActions button {
-  font-size: 14px;
-  color: #555;
-}
-
-.notificationActions button:hover {
-  color: #d10000;
-}
-
-.closeButton {
-  margin-left: 10px;
-}
-
-.notificationList {
-  overflow-y: auto;
-  flex: 1;
-}
-
-.notificationItem {
-  display: flex;
-  padding: 15px 20px;
-  border-bottom: 1px solid #f0f0f0;
-  transition: background-color 0.2s;
-  cursor: pointer;
-  position: relative;
-}
-
-.notificationItem:hover {
-  background-color: #f9f9f9;
-}
-
-.unreadNotification {
-  background-color: #f0f8ff;
-}
-
-.unreadNotification:hover {
-  background-color: #e0f0ff;
-}
-
-.notificationDot {
-  width: 10px;
-  height: 10px;
-  border-radius: 50%;
-  margin-right: 12px;
-  margin-top: 6px;
-  flex-shrink: 0;
-}
-
-.dotNew {
-  background-color: #1890ff;
-}
-
-.dotAlert {
-  background-color: #faad14;
-}
-
-.dotUpdate {
-  background-color: #52c41a;
-}
-
-.notificationContent {
-  flex: 1;
-  padding-right: 20px;
-}
-
-.notificationTitle {
-  font-weight: 600;
-  margin-bottom: 4px;
-  color: #333;
-}
-
-.notificationMessage {
-  color: #666;
-  margin-bottom: 4px;
-  font-size: 14px;
-}
-
-.notificationTime {
-  font-size: 12px;
-  color: #999;
-}
-
-.notificationActionIcon {
-  position: absolute;
-  right: 15px;
-  top: 50%;
-  transform: translateY(-50%);
-  color: #1890ff;
-  font-size: 16px;
-}
-
-.emptyNotifications {
-  padding: 30px;
-  text-align: center;
-  color: #999;
-  font-style: italic;
-}
