--- client/src/components/KitchenAdmin/MealDetails/MealForm.module.css
@@ -1,155 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Bree+Serif&family=Inter:ital,opsz,wght@0,14..32,100..900;1,14..32,100..900&family=Itim&family=Roboto+Slab:wght@100..900&family=Ubuntu:ital,wght@0,300;0,400;0,500;0,700;1,300;1,400;1,500;1,700&display=swap");
-
-
-.pageContainer {
-  min-height: 100vh;
-  background-color: rgb(244, 244, 244);
-}
-
-.languageInputsContainer {
-  display: flex;
-  gap: 8px;
-  width: 100%;
-}
-
-.languageInput {
-  flex: 1;
-}
-
-.contentWrapper {
-  padding: 24px;
-  max-width: 1100px;
-  margin: 0 auto;
-}
-
-
-.formCard {
-  border-radius: 8px;
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.08);
-  margin-top: 20px;
-}
-
-.cardTitle {
-  margin-bottom: 28px !important;
-  font-family: "Ubuntu", sans-serif;
-  /* font-family: "Roboto Slab", serif; */
-  /* text-align: center; */
-  font-size: 22px !important;
-  /* color: #d10000 !important; */
-
-}
-
-.imageContainer {
-  margin-bottom: 16px;
-}
-
-.imagePlaceholder {
-  width: 100%;
-  height: 300px;
-  border: 1px dashed #d9d9d9;
-  border-radius: 4px;
-  background-color: #fafafa;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  transition: all 0.3s;
-}
-
-.imagePlaceholder:hover {
-  border-color: rgb(230, 74, 74);
-}
-
-.uploadHint {
-  text-align: center;
-  color: #888;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  height: 100%;
-}
-
-.uploadHint p {
-  margin-top: 0;
-  font-family: "Ubuntu", sans-serif;
-}
-
-.chooseImageButton {
-  margin-bottom: 24px;
-  background-color:  white;
-  border-color:  rgb(230, 74, 74);
-  color:rgb(230, 74, 74);
-  width:200px !important;
-  margin-left: 110px;
-}
-.chooseImageButton:hover{
-  background-color:  white !important;
-  border-color:  rgb(230, 74, 74) !important;
-  color:rgb(230, 74, 74) !important;
-}
-
-.ingredientsButton {
-  background-color: rgb(230, 74, 74) !important;
-  border-color:  rgb(190, 0, 0);
-  color:white;
-  width:500px !important;
-  margin-left: 50px;
-}
-
-.ingredientsButton:hover, 
-.ingredientsButton:focus {
-  background-color: rgb(230, 74, 74) !important;
-  border-color:  rgb(230, 74, 74);
-  color: white;
-}
-
-.actionButtonsRow {
-  margin-top: 24px;
-  padding-top: 16px;
-  border-top: 1px solid #f0f0f0;
-}
-
-.cancelButtonCol {
-  text-align: left;
-  
-}
-.cancelButton{
-  border-color:  rgb(190, 0, 0) !important;
-  width:110px !important;
-  /* background-color: rgb(190, 0, 0) !important; */
-  color: rgb(190, 0, 0) !important;
-
-}
-.cancelButton:hover{
-  transform: scale(1.1);
-
-}
-
-.confirmButtonCol {
-  text-align: right;
-  
-}
-.confirmButton{
-  border-color:  rgb(190, 0, 0) !important;
-  width:110px !important;
-  background-color: rgb(190, 0, 0) !important;
-
-}
-.confirmButton:hover{
-  transform: scale(1.1);
-}
-
-/* Responsive adjustments */
-@media (max-width: 768px) {
-  .cancelButtonCol,
-  .confirmButtonCol {
-    text-align: center;
-    margin-top: 12px;
-  }
-  
-  /* Responsive navbar title */
-  .topic {
-    font-size: 24px !important;
-  }
-}
