+++ client/src/components/pages/Kitchen Admin/Report/report.jsx
@@ -1,0 +1,43 @@
+import React from "react";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import {
+  faCalendar,
+  faChartLine,
+  faBowlFood,
+} from "@fortawesome/free-solid-svg-icons";
+import NavBar from "../../../organisms/NavBar/NavBar";
+import Section from "../../../organisms/Kitchen/ReportSection/reportSection";
+
+
+
+const Report = () => {
+  return (
+    <NavBar
+    titleLines={["Meal", "Schedule", "Management"]}
+    menuItems={[
+      {
+        key: "1",
+        icon: <FontAwesomeIcon icon={faCalendar} />,
+        label: "Schedule",
+        link: "/kitchen-admin",
+      },
+      {
+        key: "2",
+        icon: <FontAwesomeIcon icon={faBowlFood} />,
+        label: "Meal",
+        link: "/kitchen-meal",
+      },
+      {
+        key: "3",
+        icon: <FontAwesomeIcon icon={faChartLine} />,
+        label: "Report & Analysis",
+        link: "/kitchen-report",
+      },
+    ]}
+    Comp={Section}
+  />
+       
+  );
+};
+
+export default Report;
