--- client/src/pages/KitchenAdmin/KitchenAdminDashBoard/kitchenAdminDash.jsx
@@ -1,13 +1,0 @@
-import React from "react";
-import Navbar from '../../../components/KitchenAdmin/KitchenNavbar/navbar';
-import Menu from "../../../components/KitchenAdmin/Calender/calender";
-
-const AnalysisDashboard = () => {
-    return (
-        <div>
-        <Navbar Comp={Menu}/>
-        </div>
-    );
-};
-
-export default AnalysisDashboard;
