--- client/src/pages/KitchenAdmin/MealPlan/mealPlan.module.css
@@ -1,209 +1,0 @@
-.container {
-    background-color: #F8F4DA;
-    min-height: 100vh;
-    width: 100vw;
-    display: flex;
-    flex-direction: column;
-    overflow-x: hidden;
-  
-  }
-
-
-  
-  /* Content Container */
-  .contentContainer {
-    padding: 20px 30px 20px 30px;
-  }
-  
-  /* Main Card Styles */
-  .mainCard {
-    background: white !important;
-    border-radius: 8px !important;
-    border: 1px solid #d9d9d9 !important;
-    
-  }
-  
-  .mainCard :global(.ant-card-head) {
-    border-bottom: none;
-    padding: 16px 24px 0 24px;
-    background: #f5f2e3;
-  }
-  
-  .mainCard :global(.ant-card-body) {
-    padding: 20px;
-  }
-  
-  .cardTitle {
-    font-size: 23px;
-    font-weight: bold;
-    color: #000;
-    padding: 2px 0;
-    
-  }
-  
-  /* Inner Card Styles */
-  .innerCard {
-    background: #f5f2e3 !important;
-    border: 1px solid #d9d9d9 !important;
-    border-radius: 8px !important;
-  }
-  
-  .innerCard :global(.ant-card-head) {
-    border-bottom: none;
-    padding: 5px 24px 0 24px;
-    background: #f5f2e3;
-  }
-  
-  .innerCard :global(.ant-card-body) {
-    padding: 5px 20px;
-  }
-  
-  .innerCardTitle {
-    font-size: 20px;
-    font-weight: bold;
-    color: #000;
-  }
-  
-  /* Search Row Styles */
-  .searchRow {
-    display: flex;
-    align-items: middle;
-    margin: 0 -8px;
-
-
-  }
-  
-  .searchInputCol {
-    flex: 0 0 calc(100% - 120px);
-    padding: 0 8px;
-
-  }
-  
-  .searchButtonCol {
-    flex: 0 0 120px;
-    padding: 0 8px;
-  }
-  
-  .searchInput {
-    width: 100%;
-    height: 40px;
-    border: 1px solid rgb(194, 194, 194) !important;
-
-  }
-  
-  .searchButton {
-    width: 100%;
-    height: 40px;
-    background: #6d071a !important;
-    border-color: #6d071a !important;
-  }
-  
-  .searchButton:hover {
-    background: #5a0616 !important;
-    border-color: #5a0616 !important;
-  }
-  
-  /* Meal List Container */
-  .mealListContainer {
-    display: flex;
-    margin-top: 16px;
-    border: 1px solid #d9d9d9;
-    border-radius: 4px;
-    padding: 10px;
-    background: white;
-    max-height:500px;
-    overflow-y: auto; /* Enables vertical scrolling */
-    /* height:500px; */
-  }
-  
-  .mealList {
-    flex: 1;
-  }
-  
-  .mealItem {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    padding: 6px 0;
-    font-size: 17px;
-  }
-  
-  .mealItemBorder {
-    border-bottom: 1px solid  rgb(194, 194, 194);
-  }
-  
-  .mealName {
-    margin-left: 8px;
-  }
-  
-  /* Optional: Customize scrollbar */
-.mealListContainer::-webkit-scrollbar {
-    width: 6px;
-  }
-  
-  .mealListContainer::-webkit-scrollbar-thumb {
-    background-color:  #660e1e;
-    border-radius: 3px;
-  }
-  
-  .mealListContainer::-webkit-scrollbar-track {
-    background-color:white;
-    /* border: 1px solid black; */
-    
-  }
-  /* Footer Button Styles */
-  .footerButtons {
-    margin-top: 16px;
-    display: flex;
-    justify-content: flex-end;
-  }
-  
-  .cancelButton {
-    border-radius: 4px;
-    height: 40px;
-    padding-left: 16px;
-    padding-right: 16px;
-    border-color: #6d071a !important;
-    color: #6d071a !important;
-    width:100px;
-
-  }
-  
-  .updateButton {
-    background: #6d071a !important;
-    border-color: #6d071a !important;
-    border-radius: 4px;
-    height: 40px;
-    padding-left: 16px;
-    padding-right: 16px;
-    width:100px;
-
-  }
-  
-  .updateButton:hover {
-    background: #5a0616 !important;
-    border-color: #5a0616 !important;
-  }
-
-  .checkbox{
-    transform: scale(1.2);
-  }
-
-  /* Target Ant Design's internal checkbox classes */
-.checkbox :global(.ant-checkbox-inner) {
-  /* border-color: #6b0f1a !important; */
-  width: 17px !important;
-  height: 17px !important;
-}
-
-.checkbox :global(.ant-checkbox-checked .ant-checkbox-inner) {
-  background-color: #6b0f1a !important;
-  border-color: #6b0f1a !important;
-}
-
-.checkbox :global(.ant-checkbox-checked .ant-checkbox-inner::after) {
-  border-color: white !important;
-  transform: rotate() scale(1.2) !important;
-}
-
-
