--- client/src/pages/KitchenAdmin/Report/report.jsx
@@ -1,17 +1,0 @@
-import React from "react";
-import Navbar from "../../../components/KitchenAdmin/KitchenNavbar/navbar";
-import Section from "../../../components/KitchenAdmin/ReportSection/reportSection";
-
-
-
-const Report = () => {
-  return (
-      <div>
-        <Navbar Comp={Section} /> 
- 
-      </div>
-       
-  );
-};
-
-export default Report;
