--- client/src/pages/KitchenAdmin/SelectIngredients/ingredient.module.css
@@ -1,117 +1,0 @@
-.container {
-    background-color:#F8F4DA;
-    min-height: 100vh;
-    width: 100vw;
-    display: flex;
-    flex-direction: column;
-    overflow-x: hidden;
-    padding:0px;
-  }
-
-
-.content {
-  padding: 10px 0px 0px 30px;
-  width:100vw;
-  margin: 0 auto;
-}
-
-.card {
-  border-radius: 8px;
-  margin-bottom: 0px;
-  padding: 40px;
-}
-
-.ingredientsSection {
-  margin-top: 20px;
-  background-color: #F8F4DA;
-  padding: 20px;
-  border: 1px solid #d7d4c4;
-  border-radius: 8px;
-
-}
-
-.searchInput {
-  margin-bottom: 15px;
-  font-size: 18px;
-}
-
-.ingredientsList {
-  padding: 15px;
-  margin-top: 10px;
-  max-height: 600px;
-  overflow-y: auto; /* Enables vertical scrolling */
-
-}
-
-.ingredientRow {
-  display: flex;
-  align-items: center;
-  margin-bottom: 10px;
-  gap: 20px; 
-  /* border-bottom: 1px slod black; */
-
-}
-
-.ingredientInput {
-  flex: 1;
-  font-size: 16px;
-}
-
-.checkbox {
-  margin: 0 ;
-  transform: scale(1.2);
-}
-
-/* Target Ant Design's internal checkbox classes*/
-.checkbox :global(.ant-checkbox-inner) {
- 
-  width: 17px !important;
-  height: 17px !important;
-} 
-
-
-
-.amountInput {
-  width: 120px;
-}
-
-.buttonContainer {
-  display: flex;
-  justify-content: flex-end;
-  margin-top: 10px;
-}
-
-.confirmButton {
-  background-color: #6b0f1a;
-  border-color: #6b0f1a;
-  width: 280px;
-  font-size: 16px;
-}
-
-.confirmButton:hover {
-  background-color: #8b1a27 !important;
-  border-color: #8b1a27;
-}
-/* Optional: Customize scrollbar */
-.ingredientsList::-webkit-scrollbar {
-  width: 6px;
-}
-.ingredientsList::-webkit-scrollbar-thumb {
-  background-color:  #660e1e;
-  border-radius: 3px;
-}
-.ingredientsList::-webkit-scrollbar-track {
-  background-color:white;
-  /* border: 1px solid black; */}
-
-  /* Custom hover and focus for input fields */
-.ingredientInput:hover :global(.ant-input),
-.amountInput:hover :global(.ant-input) {
-  border-color: #6b0f1a !important;
-}
-
-.ingredientInput:focus-within :global(.ant-input),
-.amountInput:focus-within :global(.ant-input) {
-  border-color: #6b0f1a !important;
-  box-shadow: 0 0 0 2px rgba(107, 15, 26, 0.2) !important;
-}
